refactor: balance recoil state로 구현
